pyproject.toml
[project]
name = "github-agent-proxy"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
    "types-pyyaml>=6.0.12.20250915",
]
